Refactored top-level zenml files
